refactor(index.ts): fix issue where `async` break the code
So if you add `async` to your routes function, it will break the code. Apparently if you want to `async` function with errors you have to pass the `error` to `next()` function. Which is weird by design.

This package named `express-async-errors` will fix that without changing your codes. Totally optional, but if you prefer to use `throw new Error` function you have to install this.