refactor(custom-error.ts):
This is not related to the current branch, but it just ticks me off since I have to keep rememberring that `Object.setPrototypeOf` method. So this is the solution. This is the link for reference towards the "Error" object and what's happening behind the scene https://stackoverflow.com/questions/41102060/typescript-extending-error-class