Problem: Default NATS behaviour will destroy messages(or event) once it has been received by its client.
Solution: Handle it manually by setting an option of `setManualAckMode` to true, then use `msg.ack()` to respond to the NATS Server.

This way when error happens the NATS Server will not just "fire and forget", instead it will wait and try to send the same request again for a few times.