Durable Subscription: In order to not "send all messages", we can use `setDurableName` options. However, this might not what you expect. When we `rs` (restart) the `listener` service, it will send the sent messages again.
However this is expected by the NATS Server, because once we're saving the `setDurableName` of `accounting-service` the NATS Server think that this service might be dead, so it will delete the durable subscription.

To prevent the deletion even when the client is down we have to set a "Queue Group". Hence the reason why we use the "Queue Group" again.